requirements.txt
Django>=5.0
mysqlclient
django-jazzmin
django-tinymce
gunicorn
whitenoise
dj-database-url
